+++ packages/gestalt/src/Video/__tests__/VideoPlayhead.test.js
@@ -1,0 +1,11 @@
+/* eslint-env jest */
+import React from 'react';
+import { create } from 'react-test-renderer';
+import VideoPlayhead from '../VideoPlayhead';
+
+test('VideoPlayhead', () => {
+  const tree = create(
+    <VideoPlayhead currentTime={50} duration={100} seek={() => {}} />
+  ).toJSON();
+  expect(tree).toMatchSnapshot();
+});