+++ CHANGELOG.md
@@ -11,9 +11,12 @@
 * Sticky: Expand threshold options to take string values (#166)
 * Avatar: Fall back to default letter if image does not load (#156)
 
+* Video: Add new Video component to Gestalt (#150)
+
 ### Patch
 
 * Internal: Add bundle size impact reporting (#146)
+* Docs: Fix home link (#170)
 
 </details>